Modify controller delete Event + Ajax
Modification de la methode suppression de l'événement + ajout requête Ajax pour la suppression dans la vue list

+++ src/AppBundle/Controller/backOffice/EventController.php
@@ -38,14 +38,14 @@ class EventController extends Controller
          */
         $paginator = $this->get('knp_paginator');
         $result = $paginator->paginate(
-                                $query,
-                                $request->query->getInt('page', 1),
-                                $request->query->getInt('limit', 5)
-                                );
-                dump($result);
+            $query,
+            $request->query->getInt('page', 1),
+            $request->query->getInt('limit', 5)
+        );
 
+        $deleteFormAjax = $this->createCustomForm(':EVENT_ID', 'DELETE', 'admin_event_delete');
 
-        return $this->render('@App/backOffice/listEvent.html.twig', ['Events' => $result]);
+        return $this->render('@App/backOffice/listEvent.html.twig', ['Events' => $result, 'delete_form_ajax' => $deleteFormAjax->createView()]);
     }
 
     /**
@@ -60,7 +60,6 @@ class EventController extends Controller
 
         // On récupère le formulaire
         $form = $this->createForm(EventType::class, $Event);
-
 
         // Requête envoyée par le formulaire
         $form->handleRequest($request);
@@ -125,23 +124,87 @@ class EventController extends Controller
     }
 
     /**
-     * @Route("/admin/event/delete/{id}", name="admin_event_delete", requirements={"id" = "\d+"})
+     * @Route("/admin/event/view/{id}", name="admin_event_view", requirements={"id" = "\d+"})
+     * @param $Event
+     * @return Response
+     */
+    public function viewAction(Request $request, Event $Event)
+    {
+//        return new Response('id: ' . $Event->getIdEvent() . ' titre: ' . $Event->getTitreEvent());
+//        $deleteForm = $this->createDeleteForm($Event);
+        $deleteForm = $this->createCustomForm($Event->getIdEvent(), 'DELETE', 'admin_event_delete');
+        // On rend la vue
+        return $this->render('@App/backOffice/viewEvent.html.twig', ['Event' => $Event, 'delete_form' => $deleteForm->createView()]);
+    }
+
+    /**
+     * @Route("/admin/event/delete/{id}", name="admin_event_delete")
      * @param $Event
      * @return Response
      */
     public function deleteAction(Request $request, Event $Event)
     {
 
-        // On enregistre la catégorie en BDD
+//        // On enregistre l'événement en BDD
+//        $em = $this->getDoctrine()->getManager();
+//
+//        // On supprime l'événement dans la bdd
+//        $em->remove($Event);
+//        $em->flush();
+//
+//        // Flash message
+//        $this->addFlash('success', 'Événement supprimé avec succès.');
+//
+//        return $this->redirectToRoute('admin_event_list');
+
+        // On enregistre l'événement en BDD
         $em = $this->getDoctrine()->getManager();
 
-        // On supprime l'événement dans la bdd
-        $em->remove($Event);
-        $em->flush();
+//        $form = $this->createDeleteForm($Event);
+        // On récupère tout les événements de l'Entity Event
+        $allEvents = $em->getRepository('AppBundle:Event')->findAll();
+        // Compte le nombre d'événements
+        $countEvents = count($allEvents);
+        $form = $this->createCustomForm($Event->getIdEvent(), 'DELETE', 'admin_event_delete');
+        $form->handleRequest($request);
 
-        // Flash message
-        $this->addFlash('success', 'Événement supprimé avec succès.');
+        if ($form->isSubmitted() && $form->isValid()) {
+            // Si c'est une requête Ajax (vue admin_event_list)
+            if ($request->isXmlHttpRequest()) {
+                // On supprime l'événement dans la bdd
+                $em->remove($Event);
+                $em->flush();
+                // Return d'une réponse Ajax
+                return new Response(
+                    json_encode(array('delete' => 'success', 'countEvents' => $countEvents, 'message' => 'Événement supprimé avec succès.')),
+                    200,
+                    array('Content-Type' => 'application/json')
+                );
+            }
+            // On supprime l'événement dans la bdd
+            $em->remove($Event);
+            $em->flush();
 
-        return $this->redirectToRoute('admin_event_list');
+            // Flash message
+            $this->addFlash('success', 'Événement supprimé avec succès.');
+
+            return $this->redirectToRoute('admin_event_list');
+        }
+    }
+
+    private function createDeleteForm(Event $Event)
+    {
+        return $this->createFormBuilder()
+            ->setAction($this->generateUrl('admin_event_delete', array('id' => $Event->getIdEvent())))
+            ->setMethod('DELETE')
+            ->getForm();
+    }
+
+    private function createCustomForm($id, $method, $route)
+    {
+        return $this->createFormBuilder()
+            ->setAction($this->generateUrl($route, array('id' => $id)))
+            ->setMethod($method)
+            ->getForm();
     }
 }
